Fix a rare bug in error reporting
When a new directory is first added to the target repository and
the commit does not go through, the working copy is left with files
scheduled for addition. The problem is that the checking for uncommitted
files is done after the last committed revision is read, so an attempt
is made to retrieve information on a directory scheduled for addition.
Subversion does return some information on such directories but that
information obviously does not contain a revision number.

This fix updates the code to skip directories for which there is no
revision number in the output of 'svn info'.